+++ src/lib/iTin.Utilities/iTin.Utilities.Pdf/iTin.Utilities.Pdf.Writer/ComponentModel/Operations/Items/Replace/ReplacementObject/Text/Items/Base/General/TextReplacementBase.cs
@@ -19,7 +19,6 @@ namespace iTin.Utilities.Pdf.Writer.ComponentModel.Replacement.Text
         #region public properties
 
         #region [public] (ReplaceTextOptions) ReplaceOptions: Gets or sets a value that represents replace text options
-        /// <inheritdoc />
         /// <summary>
         /// Gets or sets a value that represents text replace options.
         /// </summary>