+++ src/test/NetCore/iPdfWriter.ConsoleAppCore31/Program.cs
@@ -1,16 +1,15 @@
 ﻿
+using System;
+using iTin.Core.Models.Design.Enums;
+
+using iTin.Logging;
+using iTin.Logging.ComponentModel;
+
+using iPdfWriter.Code;
+using iPdfWriter.ComponentModel;
+
 namespace iPdfWriter
 {
-    using System;
-
-    using iTin.Logging;
-    using iTin.Logging.ComponentModel;
-
-    using iTin.Core.Models.Design.Enums;
-
-    using Code;
-    using ComponentModel;
-
     static class Program
     {
         static void Main(string[] args)
@@ -163,6 +162,12 @@ namespace iPdfWriter
             logger.Info("> Start Pdf Sample 27");
             logger.Info(" > Shows how to add a password to output file");
             Sample27.Generate(logger);
+
+            // 28. Generate sample 28 report
+            logger.Info("");
+            logger.Info("> Start Pdf Sample 28");
+            logger.Info(" > Shows how to insert an image into document");
+            Sample28.Generate(logger, YesNo.No);
 
             logger.Info("");
             logger.Debug(">End Logging<");