+++ src/test/NetCore/iPdfWriter.ConsoleAppCore60/Program.cs
@@ -162,6 +162,12 @@ namespace iPdfWriter
             logger.Info(" > Shows how to add a password to output file");
             Sample27.Generate(logger);
 
+            // 28. Generate sample 28 report
+            logger.Info("");
+            logger.Info("> Start Pdf Sample 28");
+            logger.Info(" > Shows how to insert an image into document");
+            Sample28.Generate(logger, YesNo.Yes);
+
             logger.Info("");
             logger.Debug(">End Logging<");
             Console.ReadKey();